light-controller/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.10
docopt==0.6.2
frozenlist==1.2.0
idna==3.3
multidict==5.2.0
paho-mqtt==1.6.1
requests==2.27.1
urllib3==1.26.8
yarl==1.7.2
